refactor: make List to follow compound component pattern
