+++ packages/components/list/src/index.ts
@@ -1,4 +1,4 @@
-export { List } from './List';
+export { List } from './CompoundList';
 export type { ListProps } from './List';
 
 export { ListItem } from './ListItem/ListItem';
