+++ packages/forma-36-codemod/transforms/__testfixtures__/v4-list.input.js
@@ -1,4 +1,7 @@
 import React from "react";
 import { List, ListItem } from "@contentful/forma-36-react-components";
 
-<List element="ul" className="className" testId='testId'><ListItem>ListItem1</ListItem><ListItem>ListItem2</ListItem></List>
+<List element="ul" className="className" testId='testId'>
+  <ListItem>ListItem1</ListItem>
+  <ListItem>ListItem2</ListItem>
+</List>;
