+++ packages/forma-36-codemod/transforms/__testfixtures__/v4-list.output.js
@@ -2,4 +2,7 @@ import React from "react";
 
 import { List, ListItem } from "@contentful/f36-components";
 
-<List as="ul" className="className" testId='testId'><ListItem>ListItem1</ListItem><ListItem>ListItem2</ListItem></List>
+<List as="ul" className="className" testId='testId'>
+  <ListItem>ListItem1</ListItem>
+  <ListItem>ListItem2</ListItem>
+</List>;
